fix: tests client store and int conversion

+++ internal/database/action_test.go
@@ -74,7 +74,7 @@ func TestActionRepo_Store(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 
 			// Actual test for Store
@@ -138,7 +138,7 @@ func TestActionRepo_StoreFilterActions(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 
 			// Actual test for StoreFilterActions
@@ -217,7 +217,7 @@ func TestActionRepo_FindByFilterID(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 			createdActions, err := repo.StoreFilterActions(context.Background(), int64(createdFilters[0].ID), []*domain.Action{&mockData})
 			assert.NoError(t, err)
@@ -292,7 +292,7 @@ func TestActionRepo_List(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 			createdActions, err := repo.StoreFilterActions(context.Background(), int64(createdFilters[0].ID), []*domain.Action{&mockData})
 			assert.NoError(t, err)
@@ -342,7 +342,7 @@ func TestActionRepo_Get(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 			createdActions, err := repo.StoreFilterActions(context.Background(), int64(createdFilters[0].ID), []*domain.Action{&mockData})
 			assert.NoError(t, err)
@@ -399,7 +399,7 @@ func TestActionRepo_Delete(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 			createdActions, err := repo.StoreFilterActions(context.Background(), int64(createdFilters[0].ID), []*domain.Action{&mockData})
 			assert.NoError(t, err)
@@ -453,7 +453,7 @@ func TestActionRepo_DeleteByFilterID(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 			createdActions, err := repo.StoreFilterActions(context.Background(), int64(createdFilters[0].ID), []*domain.Action{&mockData})
 			assert.NoError(t, err)
@@ -505,7 +505,7 @@ func TestActionRepo_ToggleEnabled(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, createdFilters)
 
-			mockData.ClientID = int32(mock.ID)
+			mockData.ClientID = mock.ID
 			mockData.FilterID = createdFilters[0].ID
 			mockData.Enabled = false
 			createdActions, err := repo.StoreFilterActions(context.Background(), int64(createdFilters[0].ID), []*domain.Action{&mockData})
